+++ src/webapp/src/components/analyze/AnalyzeList.vue
@@ -17,33 +17,24 @@ export default {
   computed: {
     ...mapGetters('orchestration', ['getSuccessfulPipelines']),
     ...mapGetters('plugins', ['getInstalledPlugin', 'visibleExtractors']),
-    ...mapGetters('repos', ['urlForModelDesign']),
+    ...mapGetters('repos', ['hasModels', 'urlForModelDesign']),
     ...mapState('repos', ['models']),
     contextualModels() {
-      // Default to all unless pipeline is provided for contextual display
-      let models = this.models
+      const contextualModels = {}
 
-      // Contextual dropdown based on specific pipeline
       if (this.pipeline) {
         // Split based on '@' profiles convention
         const extractor = this.pipeline.extractor.split('@')[0]
         const namespace = this.getInstalledPlugin('extractors', extractor)
           .namespace
-        const filteredModels = {}
-        for (const prop in models) {
-          if (models[prop].plugin_namespace === namespace) {
-            filteredModels[prop] = models[prop]
+        for (const prop in this.models) {
+          if (this.models[prop].plugin_namespace === namespace) {
+            contextualModels[prop] = this.models[prop]
           }
         }
-
-        // Fallback to all if no match
-        models =
-          Object.keys(filteredModels).length === 0
-            ? this.models
-            : filteredModels
       }
 
-      return models
+      return contextualModels
     },
     getIsModelEnabled() {
       return model => {
@@ -61,6 +52,15 @@ export default {
           plugin => plugin.namespace === model.plugin_namespace
         )
       }
+    },
+    getTargetModels() {
+      return this.pipeline ? this.contextualModels : this.models
+    },
+    hasContextualModels() {
+      return Object.keys(this.contextualModels).length > 0
+    },
+    isShowLackingContextualModelsMessage() {
+      return this.hasModels && this.pipeline && !this.hasContextualModels
     }
   },
   methods: {
@@ -78,31 +78,44 @@ export default {
 
 <template>
   <div>
-    <div
-      v-for="(model, modelKey) in contextualModels"
-      :key="`${modelKey}-panel`"
-      class="box box-analyze-nav is-borderless is-shadowless is-marginless"
-    >
-      <div class="content">
-        <h3 class="is-size-6">
-          {{ model.name | capitalize | underscoreToSpace }}
-        </h3>
-        <h4 class="is-size-7 has-text-grey">
-          {{ model.namespace }}
-        </h4>
-      </div>
-      <div class="buttons">
-        <router-link
-          v-for="design in model['designs']"
-          :key="design"
-          class="button is-small is-interactive-primary is-outlined"
-          :disabled="!getIsModelEnabled(model)"
-          :to="urlForModelDesign(modelKey, design)"
-          @click.native="prepareAnalyzeLoader(model.name, design)"
-          >{{ design | capitalize | underscoreToSpace }}</router-link
+    <template v-if="isShowLackingContextualModelsMessage">
+      <div class="content has-text-centered">
+        <p>There are no models for this pipeline yet.</p>
+        <a
+          href="https://www.meltano.com/docs/architecture.html#meltano-model"
+          target="_blank"
+          class="button is-small is-text"
+          >Create a Model</a
         >
       </div>
-    </div>
+    </template>
+    <template v-else>
+      <div
+        v-for="(model, modelKey) in getTargetModels"
+        :key="`${modelKey}-panel`"
+        class="box box-analyze-nav is-borderless is-shadowless is-marginless"
+      >
+        <div class="content">
+          <h3 class="is-size-6">
+            {{ model.name | capitalize | underscoreToSpace }}
+          </h3>
+          <h4 class="is-size-7 has-text-grey">
+            {{ model.namespace }}
+          </h4>
+        </div>
+        <div class="buttons">
+          <router-link
+            v-for="design in model['designs']"
+            :key="design"
+            class="button is-small is-interactive-primary is-outlined"
+            :disabled="!getIsModelEnabled(model)"
+            :to="urlForModelDesign(modelKey, design)"
+            @click.native="prepareAnalyzeLoader(model.name, design)"
+            >{{ design | capitalize | underscoreToSpace }}</router-link
+          >
+        </div>
+      </div>
+    </template>
   </div>
 </template>